requirements.txt
requests 
bs4
pandas
datetime
pymongo
ipython
psycopg2
psycopg2-binary
python-dotenv
boto3

# os
# time
# re
# random 
# warnings
# time
# IPython.core.display
# numpy
